// sanity/sanity.types.ts
export type Address = {
  _id: string
  _type: 'address'
  _createdAt: string
  _updatedAt: string
  _rev: string
  user?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'user'
  }
  name?: string
  address?: string
  city?: string
  state?: string
  pin?: string
  default?: boolean
  createdAt?: string
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol